requirements.txt
pyserial==3.2.1
python-rtmidi==1.0.0
